Fix #4 Auto locate config file if none configured
This fixes the issue where linting only works for files in the project
root directory.

If no config file is set by the user then try to find one based on the
current file being linted, instead of defaulting to `.php_cs`.

To find one, we traverse up the file path looking for it.